refactor: Centraliza lógica de API e separa tipos de constantes
