+++ src/services/transactionTypeService.ts
@@ -1,8 +1,10 @@
 import { API_CONFIG } from './api';
-import type { TransactionType } from '@/types/transactionType';
+import type { TransactionCategoryOption } from '@/types/transactionType';
 import { handleFetchError, fetchWithTimeout } from './apiUtils';
 
-export async function getTransactionTypes(): Promise<TransactionType[]> {
+export async function getTransactionTypes(): Promise<
+  TransactionCategoryOption[]
+> {
   try {
     const response = await fetchWithTimeout(
       `${API_CONFIG.BASE_URL}${API_CONFIG.ENDPOINTS.TRANSACTION_TYPES}`,
@@ -21,6 +23,6 @@ export async function getTransactionTypes(): Promise<TransactionType[]> {
     const data = await response.json();
     return data;
   } catch (error) {
-    handleFetchError(error);
+    throw handleFetchError(error);
   }
 }
