+++ src/types/newAccount.ts
@@ -1,0 +1,6 @@
+export interface NewAccountProps {
+  id: string;
+  label: string;
+  children?: NewAccountProps[];
+  isActive?: boolean;
+}
